+++ test/turbo_streamer/errors_test.rb
@@ -1,0 +1,47 @@
+require 'test_helper'
+
+class TurboStreamer::ErrorsTest < ActiveSupport::TestCase
+
+  test 'a key with no value raises' do
+    error = assert_raises(TurboStreamer::MissingValueError) do
+      jbuild do |json|
+        json.object! { json.foo }
+      end
+    end
+
+    assert_equal "No value given for `foo`.", error.message
+  end
+
+  test 'set! with no value raises' do
+    error = assert_raises(TurboStreamer::MissingValueError) do
+      jbuild do |json|
+        json.object! { json.set! :foo }
+      end
+    end
+
+    assert_equal "No value given for `foo`.", error.message
+  end
+
+  test 'child! with no value raises' do
+    error = assert_raises(TurboStreamer::MissingValueError) do
+      jbuild do |json|
+        json.array! { json.child! }
+      end
+    end
+
+    assert_equal "No value given for `child!`.", error.message
+  end
+
+  test 'a key with an explicit nil value is still allowed' do
+    assert_json({'foo' => nil}) do |json|
+      json.object! { json.foo nil }
+    end
+  end
+
+  test 'a key with false is not mistaken for a missing value' do
+    assert_json({'foo' => false}) do |json|
+      json.object! { json.foo false }
+    end
+  end
+
+end